Restructure server code, proper dev setup in package.json
During development, the server is now restarted on changes again. I'm
still not sure how to feel about including the side-effect code in
src/index.ts (where it starts the server), because that doesn't seem
right.
Then again, I couldn't think of a better structure.. having an index.js
in the root directory didn't work with ts-node, because nothing was
reloaded on changes.